+++ src/main/java/net/geforcemods/securitycraft/blocks/SentryDisguiseBlock.java
@@ -28,11 +28,6 @@ public class SentryDisguiseBlock extends DisguisableBlock {
 	}
 
 	@Override
-	public boolean isAir(BlockState state, IBlockReader world, BlockPos pos) {
-		return state.getValue(INVISIBLE);
-	}
-
-	@Override
 	public boolean propagatesSkylightDown(BlockState state, IBlockReader level, BlockPos pos) {
 		return state.getValue(INVISIBLE);
 	}